test: move unit-tests to subfolder
