+++ __tests__/integration-tests/dummy.test.ts
@@ -1,0 +1,3 @@
+test('dummy-test', () => {
+	expect(true).toBe(true);
+});
