+++ __tests__/unit-tests/MediaService.startWorkers.test.ts
@@ -5,11 +5,11 @@ import { Worker } from 'mediasoup/node/lib/Worker';
 jest.mock('mediasoup/node/lib/Worker');
 
 import 'jest';
-import MediaService, { MediaServiceOptions } from '../src/MediaService';
-import WorkerMock from '../__mocks__/WorkerMock';
+import MediaService, { MediaServiceOptions } from '../../src/MediaService';
+import WorkerMock from '../../__mocks__/WorkerMock';
 import EventEmitter from 'events';
 import { EnhancedEventEmitter } from 'mediasoup/node/lib/EnhancedEventEmitter';
-import TestUtils from '../__mocks__/TestUtils';
+import TestUtils from '../../__mocks__/TestUtils';
 
 const emptyMediaServiceOptions = {} as unknown as MediaServiceOptions;
 const optionsWithWorkers = {
