+++ __tests__/unit-tests/MediaService.test.ts
@@ -6,9 +6,9 @@ jest.mock('mediasoup/node/lib/Worker');
 jest.mock('@observertc/sfu-monitor-js');
 
 import 'jest';
-import MediaService, { MediaServiceOptions, WorkerData } from '../src/MediaService';
+import MediaService, { MediaServiceOptions, WorkerData } from '../../src/MediaService';
 import * as observeRtcMock from '@observertc/sfu-monitor-js';
-import WorkerMock from '../__mocks__/WorkerMock';
+import WorkerMock from '../../__mocks__/WorkerMock';
 import EventEmitter from 'events';
 import { EnhancedEventEmitter } from 'mediasoup/node/lib/EnhancedEventEmitter';
 import { Transport } from 'mediasoup/node/lib/Transport';
