+++ __tests__/unit-tests/RoomServerConnection.test.ts
@@ -1,6 +1,6 @@
 import { BaseConnection, Pipeline, SocketMessage } from 'edumeet-common';
-import { RoomServerConnection, RoomServerConnectionContext, RoomServerConnectionOptions } from '../src/RoomServerConnection';
-import BaseConnectionMock from '../__mocks__/BaseConnectionMock';
+import { RoomServerConnection, RoomServerConnectionContext, RoomServerConnectionOptions } from '../../src/RoomServerConnection';
+import BaseConnectionMock from '../../__mocks__/BaseConnectionMock';
 
 const createExecuteSpyAndMock = (handled: boolean) => {
 	const pipelineMock = {
