+++ __tests__/unit-tests/middlewares/producerMiddleware/producer.test.ts
@@ -3,15 +3,15 @@ import { DataProducer } from 'mediasoup/node/lib/DataProducer';
 import { Producer, ProducerScore } from 'mediasoup/node/lib/Producer';
 import { Router } from 'mediasoup/node/lib/Router';
 import { Transport } from 'mediasoup/node/lib/Transport';
-import { MiddlewareOptions } from '../../../src/common/types';
-import { createProducerMiddleware } from '../../../src/middlewares/producerMiddleware';
-import RoomServer from '../../../src/RoomServer';
-import { RoomServerConnection, RoomServerConnectionContext } from '../../../src/RoomServerConnection';
-import ProducerMock from '../../../__mocks__/ProducerMock';
-import RoomServerConnectionMock from '../../../__mocks__/RoomServerConnectionMock';
-import RoomServerMock from '../../../__mocks__/RoomServerMock';
-import RouterMock from '../../../__mocks__/RouterMock';
-import TransportMock from '../../../__mocks__/TransportMock';
+import { MiddlewareOptions } from '../../../../src/common/types';
+import { createProducerMiddleware } from '../../../../src/middlewares/producerMiddleware';
+import RoomServer from '../../../../src/RoomServer';
+import { RoomServerConnection, RoomServerConnectionContext } from '../../../../src/RoomServerConnection';
+import ProducerMock from '../../../../__mocks__/ProducerMock';
+import RoomServerConnectionMock from '../../../../__mocks__/RoomServerConnectionMock';
+import RoomServerMock from '../../../../__mocks__/RoomServerMock';
+import RouterMock from '../../../../__mocks__/RouterMock';
+import TransportMock from '../../../../__mocks__/TransportMock';
 
 const next = jest.fn();
