+++ __tests__/unit-tests/middlewares/routerMiddleWare/router.test.ts
@@ -3,18 +3,18 @@ import { Consumer } from 'mediasoup/node/lib/Consumer';
 import { Producer } from 'mediasoup/node/lib/Producer';
 import { Router } from 'mediasoup/node/lib/Router';
 import { Transport } from 'mediasoup/node/lib/Transport';
-import { MiddlewareOptions } from '../../../src/common/types';
-import MediaService from '../../../src/MediaService';
-import { createRouterMiddleware } from '../../../src/middlewares/routerMiddleware';
-import RoomServer from '../../../src/RoomServer';
-import { RoomServerConnection, RoomServerConnectionContext } from '../../../src/RoomServerConnection';
-import ConsumerMock from '../../../__mocks__/ConsumerMock';
-import MediaServiceMock from '../../../__mocks__/MediaServiceMock';
-import ProducerMock from '../../../__mocks__/ProducerMock';
-import RoomServerConnectionMock from '../../../__mocks__/RoomServerConnectionMock';
-import RoomServerMock from '../../../__mocks__/RoomServerMock';
-import RouterMock from '../../../__mocks__/RouterMock';
-import TransportMock from '../../../__mocks__/TransportMock';
+import { MiddlewareOptions } from '../../../../src/common/types';
+import MediaService from '../../../../src/MediaService';
+import { createRouterMiddleware } from '../../../../src/middlewares/routerMiddleware';
+import RoomServer from '../../../../src/RoomServer';
+import { RoomServerConnection, RoomServerConnectionContext } from '../../../../src/RoomServerConnection';
+import ConsumerMock from '../../../../__mocks__/ConsumerMock';
+import MediaServiceMock from '../../../../__mocks__/MediaServiceMock';
+import ProducerMock from '../../../../__mocks__/ProducerMock';
+import RoomServerConnectionMock from '../../../../__mocks__/RoomServerConnectionMock';
+import RoomServerMock from '../../../../__mocks__/RoomServerMock';
+import RouterMock from '../../../../__mocks__/RouterMock';
+import TransportMock from '../../../../__mocks__/TransportMock';
 
 test('getRouter() should get router', async () => {
 	const roomServer = new RoomServerMock() as unknown as RoomServer;
